uhh... trying to remove the dependency on so many similar types. Accidentally broke the parsing logic completely.
